backend/requirements.txt
Flask==3.0.0
pandas==2.2.0
numpy==1.26.4
nltk==3.8.1
spacy==3.7.2
pytesseract==0.3.10
Pillow==10.2.0
flask-cors==4.0.0
flask-caching==2.1.0
flask-limiter==3.5.0
scikit-learn==1.4.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
